--- TenByTen/TenByTen/GameUtil.h
@@ -1,5 +1,0 @@
-#pragma once
-void gotoxy(int x, int y);
-void cursor_view(bool playing);
-int key_control();
-void clear_screen();
