--- TenByTen/TenByTen/KeyDef.h
@@ -1,16 +1,0 @@
-#pragma once
-
-constexpr int board_row = 10;
-constexpr int board_col = 10;
-constexpr int block_row = 5;
-constexpr int block_col = 5;
-
-#define k_up    1
-#define k_down  2
-#define k_left  3
-#define k_right 4
-#define k_enter 5
-#define k_1     6
-#define k_2     7
-#define k_3     8
-#define k_back  9
