+++ TenByTen/TenByTen/x64/Debug/TenByTen.tlog/Cl.items.tlog
@@ -1,1 +1,1 @@
-D:\TenByTen\TenByTen\TenByTen\ngy_test.cpp;D:\TenByTen\TenByTen\TenByTen\x64\Debug\ngy_test.obj
+D:\TenByTen\TenByTen\TenByTen\ngy.cpp;D:\TenByTen\TenByTen\TenByTen\x64\Debug\ngy.obj
