+++ TenByTen/TenByTen/x64/Debug/TenByTen.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^D:\TENBYTEN\TENBYTEN\TENBYTEN\X64\DEBUG\NGY_TEST.OBJ
+^D:\TENBYTEN\TENBYTEN\TENBYTEN\X64\DEBUG\NGY.OBJ
 D:\TenByTen\TenByTen\TenByTen\x64\Debug\TenByTen.ilk
